Ajustes API e Web
Na API, a nota é int e a media é double
No blazor WEB, ajustes para mostrar classificação

+++ ScreenSound.API/Response/AvaliacaoArtistaResponse.cs
@@ -1,4 +1,4 @@
 ﻿namespace ScreenSound.API.Response
 {
-    public record AvaliacaoArtistaResponse(int artistaID, string artistaNome, int nota);
+    public record AvaliacaoArtistaResponse(int artistaID, string artistaNome, double nota);
 }
